src/requirements.txt
paho-mqtt
pycrypto
